api/requirements.txt
# Core API dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.6
boto3==1.34.34
anthropic==0.18.1
httpx==0.26.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
prometheus-client==0.19.0
structlog==24.1.0
requests==2.31.0

# Authentication dependencies
python3-saml==1.16.0
PyJWT==2.8.0
